完成对订单修改的优化

+++ hotel_management_system/src/main/java/com/patrick/controller/OrdersController.java
@@ -64,7 +64,12 @@ public class OrdersController {
     }
 
     @DeleteMapping("{oid}")
-    public RespBean delete(@PathVariable("oid") Integer oid){
+    public RespBean delete(@PathVariable("oid") Integer oid) throws MyException {
+        Integer ostate = ordersService.selectById(oid).getOstate();
+        if (ostate == 1) {
+
+            throw  new MyException("该订单已使用，不能重复使用");
+        }
         if (ordersService.delete(oid)) {
             return RespBean.ok("使用成功");
         }
